Move Table to right folder
