+++ chartlets.py/tests/components/table_test.py
@@ -1,4 +1,4 @@
-from chartlets.components.charts.table import Table
+from chartlets.components.table import Table
 from tests.component_test import make_base
